Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I disagree with the way this is implemented, although it is a great idea. Currently this will just ignore building the entire app, which is not a good idea. We can just ifneq out the Camera stuff and leave the Gallery stuff

Patch-set: 1