Update patch set 1
Patch Set 1:

I didn't design this. I only made a few tiny insignificant changes. Use this with http://review.cyanogenmod.com/#/c/25113/ and to enable the system, just add LegacyCamera to your product packages, though it won't be built because of a change in the Android.mk file in LegacyCamera

Patch-set: 1