Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

For this to work it's pretty much going to have to be reauthored. I understand these are based off of my changes to both Gallery2 and LegacyCamera but they were forks. Add in the ifdefs or ifndefs but try to keep them all in one group and not all over the place. Also the libjnimosaic_legacy (or whatever it is) is also needed and any defines (in the board config for example) need to build everything: the new Gallery2, the LegacyCamera, and the correct mosaic lib.

Patch-set: 3