Update patch set 1
Patch Set 1:

yes, it was to fix the patch you derped and got merged which broke every non legacy device..

authorship is not broken, it doesnt let you upload a new patchset to a merged commit.

As I said, I was trying to make it where I could build again, I don't need this. The other commit will likely be merged anyway.

Patch-set: 1