Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(1 inline comment)

The idea is good but when you remove the hardcoded string and make it translateable you have to add the string in strings.xml. Iy you don't the string cant't be found and that causes build errors.

Patch-set: 1