Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I totally agree with Nihanth, I always wanted something like that.

I'm not a fan of auto-rotate and I always keep it off, the only reason I turn it on is when I have to watch pics in gallery and I rotate the phone on landscape photos to see them using all the display.

Patch-set: 1