Update patch set 13
Patch Set 13:

Xaio I wasn't talking to you. the force reboot is a kernel issue,
 
You should show the steps how to reproduce the gray issue. we decided to improve surface texture patch, there may be some derp. i had a perfect solution but it leaks memory and that was the area for improvement

Patch-set: 13