Update patch set 14
Patch Set 14: Code-Review+1

I fixed the issue when camera's surface texture is null.
What happened on pause, it calls releasesurfacetexture  then it will eventually assign surfacetexture= null;
So i added a check so it won't do it again. it will create a new surface texture when it is null

Patch-set: 14