Update patch set 5
Patch Set 5:

probably dumb question but i see the fix exposure comment. I'm trying to get the pyramid (msm8660-common) going with 10.2, my last issue is that the "exposure" is set either to high to low when you open the camera. tapping the screen fixes it...this couldnt be reated to that, I think not as i dont even think or device supports HDR lol

Patch-set: 5