Update patch set 1
Patch Set 1:

Well, that's besides my point, which is: Why would we need audio effect controls specific to videos _at all_? IIRC, the philosophy behind DSPManager is to make the FX output-specific instead of app-specific. Why would account for that philosophy when playing audio (-> Apollo), but not when playing video?

Patch-set: 1