Update patch set 1
Patch Set 1:

Yeah, I agree. DSPManager is the right way to do this.

What I liked about the original patch was that it didn't interrupt the video in order to let you play with the effects. 

I think it was Kmobs that posted a concept screenshot of a neat way to add EQ settings to the volume panel. We should totally do that :)

Patch-set: 1