Update patch set 1
Patch Set 1: Code-Review+1

Devices without support for taking a snapshot while recording a video would result in a freeze and eventual FC of Gallery2. This commit properly disables the functionality for devices that opt in while allowing the functionality for supported devices. Tested on shooter, pyramid, holiday, d710, m7spr and jewel.

Patch-set: 1