Update patch set 2
Patch Set 2:

Sorry for hijacking this patch, but I want the reviewers in here to take a look at these camera screenshots.


https://docs.google.com/file/d/0B-xwGskmxvpcbVRTTmpKVDdDV2s/edit


https://dl.dropboxusercontent.com/u/16268774/Screenshots/Tablet/Screenshot_2013-09-12-00-53-53.png


https://www.dropbox.com/s/cz1jxoebv8xzypx/download_20130912_110951.jpg


As you can see, the Scene mode option is showing the wrong text above the icons on multiple devices.


My guess is, that it is correct on some devices and wrong on others, depending on how many scene modes are available for a specific devices.


In the CyanogenMod Translations G+ community, so far the following devices have been confirmed to have this problem:

-i9100

-i9300

-pollux

-p3100



I have been looking through the code, the translations are fine, but as this is also wrong in English, the problem lies deeper in the code and my knowledge of that is limited.


So please take a look at the Scene mode code. I suspect this is also a problem in AOSP, because I once flashed a apk file of the stock camera and it had the same thing.


I hope you guys can fix it for CM10.2.

Patch-set: 2