Update patch set 4
Patch Set 4:

So we cannot do anything to proprietary binaries, so (i think) photomodule is a right place for this patch on certain devices. Maybe it's possible to add a condition for this code to work on motorola only or smth?

Patch-set: 4
Reviewer: Andrey Makarov <7204@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0