Update patch set 1
Patch Set 1:

On i9100, after the recent camera commits, when taking a picture, the picture stays on the screen instead of sliding away and showing the preview again to be able to take another picture.

http://forum.xda-developers.com/showthread.php?p=46134202#post46134202

When you swipe to see the picture and go back, the preview is visible again.

Not sure it is related to this patch in particular, but it is definitely different now than before Oct 1st.

Patch-set: 1
Reviewer: Eddy Witkamp <2881@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0