Update patch set 3
Patch Set 3:

This is required for all the Samsung devices that use ZSL. I have tested this on I9300 and N7100 and this is required to show the captured image when using ImageCaptureIntent.

Contacts -> Choose a Contact -> Tap the Contact Image -> Take Photo -> Take a snapshot. The preview doesn't stop to show the captured image if ZSL is enabled. But on devices without ZSL, the preview is stopped to show the captured frame and it is restarted when Retake button is pressed. 

It is true that ZSL requires the preview to be running even during the capture, but since ImageCaptureIntent mode doesn't require taking more than 1 consecutive shot, we could disregard ZSL here.

Patch-set: 3