Update patch set 1
Patch Set 1: Code-Review-1

Josh, since this fixes a problem with a specific device and is probably not the right change for all devices, can you make this a config overlay option? If the overlay is not enabled in the device tree, it should default to FOCUS_MODE_INFINITY.

As an example of how to do this, grep packages_apps_Gallery2 for "usePreferredPreviewSizeForEffects" or "noFaceDetectOnFrontCamera".

Hit me up on IRC if you need any help.

Patch-set: 1
Reviewer: David Ferguson <3304@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1