Update patch set 1
Patch Set 1:

Because, tf700t has orientation = 0 so mOrientationResize will be true, and that will force to execute the orientation
resize code. Then an aspect ratio will be applied, so except if aspect ratio = 1, the final image size is going to be
smaller or bigger that a full screen size, but never will obtain a full screen preview. This code is only useful if
you want to show a real aspect ratio of the image of the camera, but IMO this is awful from a UI point view (I prefer
a full screen preview). Maybe, it can be added as an option.

Patch-set: 1