Update patch set 1
Patch Set 1:

There's the choice between (1) full screen preview/incorrect aspect ratio and (2) correct aspect ratio/not full screen. The patch from CodeAurora is designed to fix the fact that the aspect ratio is not correct (thus changing from 1 to 2).

We need to choose between the two, but the behavior should not differ between device and thus should not be configurable in the overlay.

Patch-set: 1