Update patch set 6
Patch Set 6:

The expanded camera menu is a problem on i9100 and probably on similar screen size devices too.

Screenshot: https://www.dropbox.com/s/315asfnv0bnbkft/Screenshot_2013-10-26-16-50-50.png

The icons on either side are cut off, but that is not the worst part. The bad part is they are non-responsive to touch!

The only way to activate the buttons on the far left and right is to swipe to them, starting from another position. But tapping them, does nothing. It does work for all the other options.

In landscape everything is fine.

Patch-set: 6
Reviewer: Eddy Witkamp <2881@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0