Update patch set 6
Patch Set 6:

My vote would be to move Storage to the extra settings menu. It is not an option that you use often, it's usually a 1 time choice.


Another thing I noticed is that the Done button on the Timer does _nothing_ anymore.
You can tap it all you want, nothing happens. Only way to close that window is to click outside of it.

Screenshot: https://www.dropbox.com/s/2exo4btdfs3smov/Screenshot_2013-10-26-17-19-54.png

Patch-set: 6