Update patch set 2
Patch Set 2:

Finally got a look at the changes. Good setuffs. Need to look at them on an N1 to see all the options, tho.

'Beautify' and 'Enhance' traditionally get a magic wand. I am partial to the one that's there, not sure where that came from? Could have some taper on the wand.

'Scene' mode selectors usually have an assortment of scene mode icons on them... Nikon arranges them on a big wheel, but we have to condense that concept to a single icon. Potentially just use one of the scene icons (that we need!) like action.

'Colorize' is generally best represented by overlapping circles of red, green and blue. Since we are monochrome, though, we can use something like the Gallery already has in the editor for it's color filters (which don't match the ones in the camera, heh). But it looks good.

'Slow shutter' is exclusively a low light mode. A night sky with moon and a few stars is a very traditional treatment.

Patch-set: 2