Update patch set 1
Patch Set 1:

Steve, Pawit, could you reconsider this change. The overlay doesn't hurt and allow the device mantainer to decide what is the best option for his users. IMO, having a 1920x1200 screen and just a preview that is the half of the screen is not a good use case.

Patch-set: 1