Update patch set 1
Patch Set 1:

I have don't know why they're only implementing it for only certain orientations, but I think it's just an incomplete orientation (90 and 270 degree would require a width/height swap)

I'm not sure what you mean by aspect ratio of "1". None of the devices have an aspect ratio of exactly 1. I didn't check the code, but I believe it doesn't correct it because it is already correct, so the code to correct it simply produce the same result as if it weren't ran. (i.e. the picture is not stretched in the view finder, although it is cropped.)

Patch-set: 1