Update patch set 4
Patch Set 4:

For me, a toast message like "Use volume keys to take a picture" (or "... to fire the shutter", if we need more general message) / "Zoom using volume keys" on option change would be good.

With the camera button we can remove these options completely — it is an expected behaviour in this case.

Patch-set: 4