Update patch set 7
Patch Set 7:

Henry: Some comments/nitpicks on the drawables:

ic_tint (hdpi): Suffers from that "the edges of a circle close to the edge is drawn out, causing it to not be completely round" problem that is quote common in the drawables Google release for developers too. Not noticable in xhdpi though.

ic_slowshutter: The drawable might look less busy if you rotate the moon a bit clock-wise. Sort of like the old Quiet Hours QS tile but flipped like yours. Might also make it ever so slightly clearer on low-res screens. I don't know if that is a problem with the current one though.

https://github.com/CyanogenMod/android_frameworks_base/commit/caa3ffb29881a0c6a11e20eadecdd5e5e76a1bb0

ic_indicator_slowshutter: Same suggestion as above.

ic_beautify: How does it look on the device? I have a feeling it might stick out.

Patch-set: 7