Update patch set 11
Patch Set 11:

Looking good. 

While we are being so artsy and creative, we also need an icon for Best Shot (or Full Auto) mode. This is currently listed as a scene mode, but I'd like to elevate it up to this menu since it's really powerful. In this mode, the sensor runs it's automatic scene detection algorithm, and adjust for various scenes such as outdoor, mixed lighting, night, landscape, etc.

I'll get the bugs with mutual exclusion of these options done this weekend.

Patch-set: 11