Update patch set 2
Patch Set 2:

I did some research and from what I found:
According to the exif spec the string has to be zero terminated with the zero included in the length. So the case where the zero is not included is actually a violation of that spec.
But it is most common to add the terminating zero instead of overwriting the last character. Seen in e.g. Windows itself (details of a jpg image) and the exiv2 parser from http://www.exiv2.org.
One could argue if it is ok to increase the length variable because another app may want to modify the value and could overwrite the first byte of the following data thinking it belongs to this tag. So I'd just add the zero but do not extend the length.
Agreed?

Patch-set: 2