Update patch set 4
Patch Set 4: Code-Review+1

That is not completely true. You get 2 times right, but only one is the recorded time. The other (only displayed as "time") is the creation or modification time of the file (not sure which one it was)
And I can certainly say that if you e.g. copy the files from another phone (maybe even just from computer or modify it) the "time" will be something newer then the record time (the latter is read from exif)

And yes this is good to go from my side.

Patch-set: 4