Gallery2: Fix UI display wrongly after deleting some pictures in photopage.
The kind in TimeLinePage is set to wrong value after retrun from photopage,
so the UI display wrongly.
The albums from dataManger is not cleared after delete all the photos,
so can't finish the photopage after deleting the last photo.

Set the right value to kind in timelinepage after return from photopage.
And set the empty path to the albums which don't exist from dataManger,
to finsih the photopage after delete the last photo.

Change-Id: I09d4d5349ce5e78a2be3129510d822584bfa57db
CRs-Fixed: 991313