Gallery2: Fix menu icon is missing when change language on Videos page.
If the gallery is in background ,the video page is saved to the stack.
After the language change, the Gallery is destroyed.
If gallery start again, the video page is restored from the stack.
But the action bar is set to don't display the home as up in onResume,
which set the navigation icon to null, so the menu icon can't display.

If albumpage is started by Video option, the action bar set to display home
option. If albumpage is start by others(albumnsetpage), the action bar set
to display home as up option,to display menu icon or up icon rightly
in action bar.
Besides, fix to display the right title with the different language.

Change-Id: Ieec0bd47400f6c0b698f5c2e842dc1038d9addcf
CRs-Fixed: 1002171