Gallery2: Fix white gallery screen before showing photo
Put a white cover View in front of the SurfaceView and hide it
after the first draw. This prevents the SurfaceView being transparent
before the first draw. So can see the white screen at first.

Set the cover View to the black color if start gallery with view single
photo action, which can match the background of photopage
to avoid see the white screen.

Change-Id: I0b3766eb6023f9bcdb7e0329dcafb978b3dc1557
CRs-Fixed: 1021668