Fix the issue of back to original.
The button back to original may not work because the process of
applying effect such as Zoom in DDM is a little slow, the activity
ImageShow can't update in time. Move the invalidateViews to the
end to make sure the update can happen in time and remove the
unnecessary actions.

Change-Id: I3a138473047babba12d649f41a31e86416f47e03
CRs-Fixed: 2087036