Gallery2: Preserve the decimals in coordinates
Currently the first index of the Rational variable is extracted and rest is just discarded. This caused the location tag to lose its precision and show incorrect values. This also caused reverse lookup to fail.

Let's just extract the values at three indexes and try to get a location coordinate.

Change-Id: I21d21d2b5c62d0e6cf1a1cf73b62a947513f09d2