Gallery: Keep the decimals in coordinates
Currently the first index of the Rational variable is extracted and rest is just discarded. This caused the location tag to lose its precision and show incorrect values. This also caused reverse lookup to fail also

Let's just extract the values at three indexes and try to get a coordinate, if this fails just fallback to providing the values at first index position

Change-Id: I21d21d2b5c62d0e6cf1a1cf73b62a947513f09d2