Gallery2: Set content URI as data in share intents
Edit action expects content URI in data otherwise leads to crash due to NPE.
In case of an array of images, pick the first image for the edit action.

This is similar to what I7a1254e579cea38fd70fc13da81e1dae98f5f968 is doing with
screenshot activity in fw/base.

Signed-off-by: Aayush Gupta <aayushgupta219@gmail.com>
Change-Id: Ifdded8862e4e08dfcb05391c1cb40cf1f4f7dbd8